+++ new_pom/pages/forms/createFormPage.js
@@ -1,32 +1,16 @@
+import CreateUpdateFormComponent from "../../components/createUpdateFormComponent";
+import TableComponent from "../../components/tableComponent";
+import ToastComponent from "../../components/toastComponent";
 import FileUploaderComponent from "../../components/fileUploaderComponent";
-import { step } from "allure-js-commons";
 
 export default class CreateFormPage {
     constructor(page) {
         this.page = page;
+
+        this.createUpdate = new CreateUpdateFormComponent(this.page)
         this.fileUploader = new FileUploaderComponent(this.page);
-
-        this.formNameField = this.page.getByPlaceholder('A form name to identify your');
-        this.optionalMessageField = this.page.getByPlaceholder('Add an optional message for');
-        this.fillTemplateBtn = this.page.getByRole('button', { name: 'Fill Template' });
+        this.table = new TableComponent(this.page); 
+        this.toast = new ToastComponent(this.page);
     }
-
-    async fillFormNameField(name) {
-        await step('Fill in Form Name', async () => {
-        await this.formNameField.fill(name);
-        });
-    }
-
-    async fillOptionalMessageField(message) {
-        await step('Fill in Optional Message', async () => {
-        await this.optionalMessageField.fill(message);
-        });
-    }
-
-    async clickFillTemplateBtn() {
-        await step('Click on the "Fill Template" button', async () => {
-        await this.fillTemplateBtn.waitFor();
-        await this.fillTemplateBtn.click();
-        });
-    }
+    
 }
